+++ Src/Library/Endpoint/Endpoint.Validation.cs
@@ -25,12 +25,21 @@ public abstract partial class Endpoint<TRequest, TResponse> : BaseEndpoint, IVal
     public bool ValidationFailed => ValidationFailures.ValidationFailed();
 
     ///<inheritdoc/>
+    public void AddError(ValidationFailure failure)
+        => ValidationFailures.AddError(failure);
+
+    ///<inheritdoc/>
     public void AddError(string message, string? errorCode = null, Severity severity = Severity.Error)
         => ValidationFailures.AddError(message, errorCode, severity);
 
     ///<inheritdoc/>
     public void AddError(Expression<Func<TRequest, object>> property, string errorMessage, string? errorCode = null, Severity severity = Severity.Error)
         => ValidationFailures.AddError(property, errorMessage, errorCode, severity);
+
+    ///<inheritdoc/>
+    [DoesNotReturn]
+    public void ThrowError(ValidationFailure failure)
+        => ValidationFailures.ThrowError(failure);
 
     ///<inheritdoc/>
     [DoesNotReturn]